metaeval/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
ca-certificates=2023.7.22=hbcca054_0
dtw-python=1.3.0=pypi_0
h5py=3.10.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
joblib=1.3.2=pyhd8ed1ab_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=5_h92ddd45_netlib
libstdcxx-ng=11.2.0=h1234567_1
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
ncurses=6.4=h6a678d5_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.24.4=pypi_0
openssl=3.0.11=h7f8727e_2
pandas=2.0.3=py39h1128e8f_0
pip=23.2.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2023.3.post1=py39h06a4308_0
readline=8.2=h5eee18b_0
scikit-learn=1.0.2=py39h4dfa638_0
scipy=1.8.1=py39he49c0e8_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.12=h1ccaba5_0
tzdata=2023c=h04d1e81_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
